Actions: Don't listen for stow events
This was required in the past to trigger doze when pulling out the
device from a pocket (we were handling the "pick up" action with Flat Up sensor).
Currently Glance sensor is being used which gets triggered in the previously described case itself
thus there's no need to keep stow implementation alive.

based on @Alberto97 commit sha-id=41fb026 for LineageActions

Signed-off-by: chirayu19091997 <chirayu.agrawal4000@gmail.com>
Co-authored-by: chirayu19091997 <chirayu.agrawal4000@gmail.com>